+++ src/aqua_diagnostics/core/diagnostic.py
@@ -1,11 +1,12 @@
 import os
 import xarray as xr
 from aqua import Reader
+from aqua.exceptions import NotEnoughDataError
 from aqua.logger import log_configure
 from aqua.util import ConfigPath
 from aqua.util import load_yaml, convert_units
-from aqua.util import area_selection
-from aqua.util import DEFAULT_REALIZATION
+from aqua.util import xarray_to_pandas_freq, pandas_freq_to_string
+from aqua.util import area_selection, DEFAULT_REALIZATION
 from .output_saver import OutputSaver
 
 
@@ -47,13 +48,15 @@ class Diagnostic():
         # Data to be retrieved
         self.data = None
 
-    def retrieve(self, var: str | None = None, reader_kwargs: dict = {}):
+    def retrieve(self, var: str | None = None, reader_kwargs: dict = {},
+                 months_required: int | None = None):
         """
         Retrieve the data from the model.
 
         Args:
             var (str | None): The variable to be retrieved. If None, all variables will be retrieved.
             reader_kwargs (dict): Additional keyword arguments to be passed to the Reader.
+            months_required (int | None): The number of months of data required. If None, no check will be performed.
 
         Attributes:
             self.data: The data retrieved from the model. If return_data is True, the data will be returned.
@@ -62,9 +65,11 @@ class Diagnostic():
         self.data, self.reader, self.catalog = self._retrieve(model=self.model, exp=self.exp, source=self.source,
                                                               var=var, catalog=self.catalog, startdate=self.startdate,
                                                               enddate=self.enddate, regrid=self.regrid,
-                                                              loglevel=self.loglevel, reader_kwargs=reader_kwargs)
+                                                              reader_kwargs=reader_kwargs, months_required=months_required,
+                                                              loglevel=self.logger.level)
 
         self.realization = self.reader.kwargs['realization'] if 'realization' in reader_kwargs else DEFAULT_REALIZATION
+
         if self.regrid is not None:
             self.logger.info(f'Regridded data to {self.regrid} grid')
         if self.startdate is None:
@@ -101,6 +106,7 @@ class Diagnostic():
                                   realization=self.realization,
                                   outputdir=outputdir, loglevel=self.loglevel)
 
+
         outputsaver.save_netcdf(dataset=data, diagnostic_product=diagnostic_product, rebuild=rebuild,
                                 create_catalog_entry=create_catalog_entry, dict_catalog_entry=dict_catalog_entry,
                                 **kwargs)
@@ -108,6 +114,7 @@ class Diagnostic():
     @staticmethod
     def _retrieve(model: str, exp: str, source: str, var: str = None, catalog: str = None,
                   startdate: str = None, enddate: str = None, regrid: str = None,
+                  months_required: int | None = None,
                   reader_kwargs: dict = {}, loglevel: str = 'WARNING'):
         """
         Static method to retrieve data and return everything instead of updating class
@@ -124,6 +131,7 @@ class Diagnostic():
             enddate (str): The end date of the data to be retrieved.
                            If None, all available data will be retrieved.
             regrid (str): The target grid to be used for regridding. If None, no regridding will be done.
+            months_required (int or None): The minimal amount of months to have results. If they are not met, a NotEnoughDataError will be raised.
             reader_kwargs (dict): Additional keyword arguments to be passed to the Reader.
             loglevel (str): The log level to be used. Default is 'WARNING'.
 
@@ -141,6 +149,26 @@ class Diagnostic():
         # If the data is empty, raise an error
         if not data:
             raise ValueError(f"No data found for {model} {exp} {source} with variable {var}")
+
+        # If there is a month requirement we infer the data frequency,
+        # then we check how many months are available in the data
+        # and finally raise an error if the requirement is not met.
+        if months_required is not None:
+            timedelta = xarray_to_pandas_freq(data)
+            freq = pandas_freq_to_string(timedelta)
+            factor = {
+                'hourly': 1/(24*30),
+                'daily': 1/30,
+                'weekly': 1/4,
+                'monthly': 1,
+                'seasonal': 3,
+                'annual': 12
+            }
+            # We automatically raise an error if the frequency is not pandas compliant
+            months = len(data['time']) * factor.get(freq, 0)
+
+            if months < months_required:
+                raise NotEnoughDataError(f"Not enough months of data found for {model} {exp} {source}, at least {months_required} months required, only {months} found.")
 
         if catalog is None:
             catalog = reader.catalog