+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Workflow modifications:
 
 AQUA core complete list:
 - Split out plotting function for vertical profile and add contour option (#2190)
+- Refactor of some `aqua.util.time` function, improving name and pandas integration (#2205)
 - Refactor of the `dump_yaml` utility function, now correctly handling `None` values as `null` (#2198)
 - `Reader` will now turn off areas and grids capabilities when `src_grid_name` is `False` (#2198)
 - LRA and `OutputSaver` jinja-related duplicated methods are now merged (#2198)
@@ -36,6 +37,7 @@ AQUA core complete list:
 - `apply_circular_window()` utility function to apply a circular window to cartopy maps (#2100)
 
 AQUA diagnostics complete list:
+- Diagnostic core: `retrieve` and `_retrieve` methods can take a `months_required` argument so that diagnostics can raise an error if insufficient months of data are available. (#2205)
 - Timeseries: introduction of the catalog entry capability, default in CLI (#2198)
 - Diagnostic core: introduction of the catalog entry capability and `self.realization` attribute (#2198)
 - Ensemble: Updating the ensemble module according the the issue #1925 (#2004)